added unit- and e2e-tests for revamped Passwd()
