+++ internal/pkg/util/fs/files/files_linux_test.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -42,35 +42,6 @@ func TestGroup(t *testing.T) {
 	f.Close()
 
 	_, err = Group(emptyGroup, uid, gids, nil)
-	if err != nil {
-		t.Error(err)
-	}
-}
-
-func TestPasswd(t *testing.T) {
-	test.DropPrivilege(t)
-	defer test.ResetPrivilege(t)
-
-	uid := os.Getuid()
-
-	_, err := Passwd("/fake", "/fake", uid, nil)
-	if err == nil {
-		t.Errorf("should have failed with bad passwd file")
-	}
-	_, err = Passwd("/etc/passwd", "/home", uid, nil)
-	if err != nil {
-		t.Errorf("should have passed with correct passwd file")
-	}
-	// with an empty file
-	f, err := os.CreateTemp("", "empty-passwd-")
-	if err != nil {
-		t.Error(err)
-	}
-	emptyPasswd := f.Name()
-	defer os.Remove(emptyPasswd)
-	f.Close()
-
-	_, err = Passwd(emptyPasswd, "/home", uid, nil)
 	if err != nil {
 		t.Error(err)
 	}
